requirements.txt
category-encoders==2.4.0
coverage==6.3.2
matplotlib==3.5.1
mlflow==1.24.0
mlxtend==0.19.0
numpy==1.21.5
pandas==1.3.5
plotly==5.6.0
pytest==7.0.1
seaborn==0.11.2
scikit-learn==1.0.2
scipy==1.7.3
statsmodels==0.13.2
tqdm==4.63.0
yellowbrick==1.4
vega_datasets==0.9.0
altair==4.2.0
pandas-profiling==3.1.0
streamlit-pandas-profiling==0.1.3
interpret-community==0.24.2
lime==0.2.0.1
lightgbm==3.3.2
great-expectations==0.14.9
umap-learn==0.5.2
pingouin==0.5.1
pyspark==3.2.1
fastapi==0.75.0
uvicorn==0.17.5